+++ train.py
@@ -340,7 +340,7 @@ class SentenceClassification:
 
         sent_reps = np.vstack(sent_reps)
 
-        sent_reps = torch.from_numpy(sent_reps).to(self.device).long()
+        sent_reps = torch.from_numpy(sent_reps).to(self.device)
 
         return sent_reps
